backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.9,<1.0.0

# Database
sqlalchemy>=2.0.25,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
alembic>=1.13.0,<2.0.0

# Security
argon2-cffi>=23.1.0,<24.0.0
cryptography>=42.0.0,<44.0.0

# Validation and settings
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Rate limiting
slowapi>=0.1.9,<1.0.0

# Scheduler
apscheduler>=3.10.0,<4.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
httpx>=0.26.0,<1.0.0

# Dev tools
ruff>=0.2.0,<1.0.0
